requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
ImageHash==4.3.1
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
pyperclip==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
easyocr==1.7.1
filelock==3.14.0
fsspec==2024.5.0
imageio==2.34.1
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
shapely==2.0.4
six==1.16.0
sympy==1.12.1
threadpoolctl==3.5.0
tifffile==2024.5.22
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
